Javadoc completed, code clean-up and classes moved to matching package
