+++ jgrapht-core/src/main/java/org/jgrapht/alg/matching/DualUpdater.java
@@ -15,24 +15,57 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg.blossom;
+package org.jgrapht.alg.matching;
 
 import org.jgrapht.util.FibonacciHeap;
 
-import static org.jgrapht.alg.blossom.KolmogorovMinimumWeightPerfectMatching.EPS;
-import static org.jgrapht.alg.blossom.KolmogorovMinimumWeightPerfectMatching.INFINITY;
+import static java.lang.Math.abs;
+import static org.jgrapht.alg.matching.KolmogorovMinimumWeightPerfectMatching.*;
 
+/**
+ * This class is used by {@link KolmogorovMinimumWeightPerfectMatching} to perform dual updates, thus creating
+ * increasing the dual objective function value and creating new tight edges.
+ *
+ * @param <V> the graph vertex type
+ * @param <E> the graph edge type
+ * @author Timofey Chudakov
+ * @see PrimalUpdater
+ * @see KolmogorovMinimumWeightPerfectMatching
+ * @since June 2018
+ */
 class DualUpdater<V, E> {
+    /**
+     * Store of the information needed for the algorithm
+     */
     private State<V, E> state;
+    /**
+     * Instance of {@link PrimalUpdater} for performing immediate augmentations after dual
+     * updates when they are applicable. These speeds the overall algorithm up.
+     */
     private PrimalUpdater primalUpdater;
 
+    /**
+     * Creates a new instance of the DualUpdater
+     *
+     * @param state         the state common to {@link PrimalUpdater}, {@link DualUpdater} and {@link KolmogorovMinimumWeightPerfectMatching}
+     * @param primalUpdater primal updater used by the algorithm
+     */
     public DualUpdater(State<V, E> state, PrimalUpdater primalUpdater) {
         this.state = state;
         this.primalUpdater = primalUpdater;
     }
 
-    double updateDuals(DualUpdateType type) {
-        if (state.options.verbose) {
+    /**
+     * Method for general dual update. It operates on the whole graph and according to the strategy
+     * defined by {@code strategy} performs dual update
+     *
+     * @param type the strategy to use in dual update
+     * @return the sum of all changes of dual variables of the trees
+     */
+    public double updateDuals(DualUpdateStrategy type) {
+        long start = System.nanoTime();
+
+        if (DEBUG) {
             System.out.println("Start updating duals");
         }
         // going through all trees roots
@@ -41,29 +74,42 @@ class DualUpdater<V, E> {
             double eps = getEps(tree);
             tree.accumulatedEps = eps - tree.eps;
         }
-        if (type == DualUpdateType.MULTIPLE_TREE_FIXED_DELTA) {
+        if (type == DualUpdateStrategy.MULTIPLE_TREE_FIXED_DELTA) {
             multipleTreeFixedDelta();
-        } else if (type == DualUpdateType.MULTIPLE_TREE_CONNECTED_COMPONENTS) {
+        } else if (type == DualUpdateStrategy.MULTIPLE_TREE_CONNECTED_COMPONENTS) {
             updateDualsConnectedComponents();
         }
 
         double dualChange = 0;
         // updating trees.eps with respect to the accumulated eps
         for (Node root = state.nodes[state.nodeNum].treeSiblingNext; root != null; root = root.treeSiblingNext) {
-            if (root.tree.accumulatedEps > 0) {
+            if (abs(root.tree.accumulatedEps) > EPS) {
                 dualChange += root.tree.accumulatedEps;
                 root.tree.eps += root.tree.accumulatedEps;
             }
         }
-        if (state.options.verbose) {
+        if (DEBUG) {
             for (Node root = state.nodes[state.nodeNum].treeSiblingNext; root != null; root = root.treeSiblingNext) {
                 System.out.println("Updating duals: now eps of " + root.tree + " is " + (root.tree.eps));
             }
         }
+
+        state.statistics.dualUpdatesTime += System.nanoTime() - start;
         return dualChange;
     }
 
-    double getEps(Tree tree) {
+    /**
+     * Computes and returns the value by which the dual variables of the "+" nodes of {@code tree} can be increased
+     * and the dual variables of the "-" nodes of {@code tree} can be decreased. This value is bounded
+     * by constraints on (+, +) in-tree edges, "-" blossoms and (+, inf) edges of the {@code tree}. As the result of
+     * the lazy delta spreading technique, this value already contains the value of tree.eps.The computed
+     * value can violate the constrains on the cross-tree edges and can be equal to
+     * {@link KolmogorovMinimumWeightPerfectMatching#INFINITY}.
+     *
+     * @param tree the tree to process
+     * @return a value which can be safely assigned to tree.eps
+     */
+    private double getEps(Tree tree) {
         double eps = INFINITY;
         Edge varEdge;
         // checking minimum slack of the plus-infinity edges
@@ -85,11 +131,20 @@ class DualUpdater<V, E> {
         return eps;
     }
 
-    boolean updateDualsSingle(Tree tree) {
-        double eps = getEps(tree);
-        double eps_augment = INFINITY;
+    /**
+     * Updates the duals of the single tree. This method operates locally on a single tree. It also finds
+     * a cross-tree (+, +) edge of minimum slack and performs an augmentation if it is possible.
+     *
+     * @param tree the tree to update duals of
+     * @return true iff some progress was made and there was no augmentation performed, false otherwise
+     */
+    public boolean updateDualsSingle(Tree tree) {
+        long start = System.nanoTime();
+
+        double eps = getEps(tree);  // include only constraints on (+,+) in-tree edges, (+, inf) edges and "-' blossoms
+        double eps_augment = INFINITY; // takes into account constraints of the cross-tree edges
+        Edge augmentEdge = null; // the (+, +) cross-tree edge of minimum slack
         Edge varEdge;
-        Edge augmentEdge = null;
         double delta = 0;
         for (Tree.TreeEdgeIterator iterator = tree.treeEdgeIterator(); iterator.hasNext(); ) {
             TreeEdge treeEdge = iterator.next();
@@ -106,22 +161,35 @@ class DualUpdater<V, E> {
         if (eps > eps_augment) {
             eps = eps_augment;
         }
+        // now eps takes into account all the constraints
+        if (eps > NO_PERFECT_MATCHING_THRESHOLD) {
+            throw new IllegalArgumentException(NO_PERFECT_MATCHING);
+        }
         if (eps > tree.eps) {
             delta = eps - tree.eps;
             tree.eps = eps;
-            if (state.options.verbose) {
+            if (DEBUG) {
                 System.out.println("Updating duals: now eps of " + tree + " is " + eps);
             }
         }
+
+        state.statistics.dualUpdatesTime += System.nanoTime() - start;
+
         if (augmentEdge != null && eps_augment <= tree.eps) {
             primalUpdater.augment(augmentEdge);
-            return false;
+            return false; // can't proceed with the same tree
         } else {
             return delta > EPS;
         }
     }
 
-    boolean updateDualsConnectedComponents() {
+    /**
+     * Updates the duals via connected components. The connect components is a set of trees which
+     * are connected via tight (+, -) cross tree edges. For these components the same dual change is
+     * chosen. As a result, the circular constrains are avoided for sure. This is the point where
+     * the {@link DualUpdater#multipleTreeFixedDelta()} approach can fail.
+     */
+    private void updateDualsConnectedComponents() {
         Node root;
         Tree startTree;
         Tree currentTree;
@@ -211,6 +279,11 @@ class DualUpdater<V, E> {
                 currentTree = currentTree.nextTree;
             }
 
+            if (eps > NO_PERFECT_MATCHING_THRESHOLD) {
+                throw new IllegalArgumentException(NO_PERFECT_MATCHING);
+            }
+
+            // applying dual change to all trees in the connected component
             Tree nextTree = startTree;
             do {
                 currentTree = nextTree;
@@ -219,13 +292,17 @@ class DualUpdater<V, E> {
                 currentTree.accumulatedEps = eps;
             } while (currentTree != nextTree);
         }
-        return false;
     }
 
-    void multipleTreeFixedDelta() {
-        if (state.options.verbose) {
+    /**
+     * Updates duals by iterating through trees and greedily increasing their dual variables. This approach
+     * can fail if there are circular constraints on (+, -) cross-tree edges.
+     */
+    private void multipleTreeFixedDelta() {
+        if (DEBUG) {
             System.out.println("Multiple tree fixed delta approach");
         }
+        Edge varEdge;
         double eps = INFINITY;
         for (Node root = state.nodes[state.nodeNum].treeSiblingNext; root != null; root = root.treeSiblingNext) {
             Tree tree = root.tree;
@@ -233,24 +310,31 @@ class DualUpdater<V, E> {
             if (eps > tree.accumulatedEps) {
                 eps = tree.accumulatedEps;
             }
+            // iterating only through outgoing tree edges so that every edge is considered only once
             for (TreeEdge outgoingTreeEdge = tree.first[0]; outgoingTreeEdge != null; outgoingTreeEdge = outgoingTreeEdge.next[0]) {
+                // since all epsilons are equal we don't have to check (+, -) cross tree edges
                 if (!outgoingTreeEdge.plusPlusEdges.isEmpty()) {
-                    Edge minEdge = outgoingTreeEdge.plusPlusEdges.min().getData();
+                    varEdge = outgoingTreeEdge.plusPlusEdges.min().getData();
                     double oppositeTreeEps = outgoingTreeEdge.head[0].eps;
-                    if (2 * eps > minEdge.slack - treeEps - oppositeTreeEps) {
-                        eps = (minEdge.slack - treeEps - oppositeTreeEps) / 2;
+                    if (2 * eps > varEdge.slack - treeEps - oppositeTreeEps) {
+                        eps = (varEdge.slack - treeEps - oppositeTreeEps) / 2;
                     }
                 }
             }
+        }
+        if (eps > NO_PERFECT_MATCHING_THRESHOLD) {
+            throw new IllegalArgumentException(NO_PERFECT_MATCHING);
         }
         for (Node root = state.nodes[state.nodeNum].treeSiblingNext; root != null; root = root.treeSiblingNext) {
             root.tree.accumulatedEps = eps;
         }
     }
 
-    enum DualUpdateType {
+    /**
+     * Enum for choosing dual update strategy
+     */
+    enum DualUpdateStrategy {
         MULTIPLE_TREE_FIXED_DELTA,
         MULTIPLE_TREE_CONNECTED_COMPONENTS,
-        MULTIPLE_TREE_STRONGLY_CONNECTED_COMPONENTS,
     }
 }
