+++ jgrapht-core/src/test/java/org/jgrapht/alg/matching/BlossomVDebugger.java
@@ -15,8 +15,9 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg.blossom;
+package org.jgrapht.alg.matching;
 
+import org.jgrapht.alg.matching.*;
 import org.jgrapht.util.FibonacciHeap;
 
 import java.util.HashSet;
