+++ jgrapht-core/src/test/java/org/jgrapht/alg/matching/InitializerTest.java
@@ -15,7 +15,7 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg.blossom;
+package org.jgrapht.alg.matching;
 
 import org.jgrapht.Graph;
 import org.jgrapht.Graphs;
@@ -25,9 +25,9 @@ import org.junit.Test;
 
 import java.util.*;
 
-import static org.jgrapht.alg.blossom.KolmogorovMinimumWeightPerfectMatching.EPS;
-import static org.jgrapht.alg.blossom.Initializer.InitializationType.GREEDY;
-import static org.jgrapht.alg.blossom.Initializer.InitializationType.NONE;
+import static org.jgrapht.alg.matching.KolmogorovMinimumWeightPerfectMatching.EPS;
+import static org.jgrapht.alg.matching.Initializer.InitializationType.GREEDY;
+import static org.jgrapht.alg.matching.Initializer.InitializationType.NONE;
 import static org.junit.Assert.*;
 
 public class InitializerTest {
