+++ jgrapht-core/src/test/java/org/jgrapht/alg/matching/StateTest.java
@@ -15,7 +15,7 @@
  * (b) the terms of the Eclipse Public License v1.0 as published by
  * the Eclipse Foundation.
  */
-package org.jgrapht.alg.blossom;
+package org.jgrapht.alg.matching;
 
 import org.jgrapht.Graph;
 import org.jgrapht.Graphs;
@@ -27,7 +27,7 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
 
-import static org.jgrapht.alg.blossom.Initializer.InitializationType.NONE;
+import static org.jgrapht.alg.matching.Initializer.InitializationType.NONE;
 import static org.junit.Assert.assertEquals;
 
 public class StateTest {
